+++ packages/terra-compact-interactive-list/src/utils/keyHandlerUtils.js
@@ -281,7 +281,7 @@ export const getFocusedCellIndexes = (list, columns, ids) => {
   const { rowId, columnId } = ids;
   const row = [...list.children].findIndex(rowElement => rowElement.getAttribute('data-row-id') === rowId);
   const cell = columns.findIndex(col => col.id === columnId);
-  // row - 1 needs to accomodate for the hidden header row
+  // row - 1 needs to accommodate for the hidden header row
   return { row: row - 1, cell };
 };
 
@@ -294,7 +294,7 @@ export const getFocusedCellIndexes = (list, columns, ids) => {
   */
 export const getFocusedCellIds = (list, columns, indexes) => {
   const { row, cell } = indexes;
-  // row + 1 needs to accomodate for the hidden header row
+  // row + 1 needs to accommodate for the hidden header row
   const rowId = list.children[row + 1].getAttribute('data-row-id');
   const columnId = columns[cell].id;
   return { rowId, columnId };