+++ packages/terra-table/src/utils/actionsUtils.js
@@ -1,0 +1,8 @@
+/**
+ * The method that checks if at least one of the columns has actions object with two parameters - text label and onCall method.
+ * @param {PropTypes.arrayOf(columnShape)} columns - array of data grid columns
+ * @returns bool
+ */
+const hasColumnActions = (columns) => (!!columns.find(column => column.action));
+
+export default hasColumnActions;