+++ packages/terra-table/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+## 5.9.0 - (February 28, 2024)
+
+* Added
+  * Column actions added.
+
 ## 5.8.0 - (February 28, 2024)
 
 * Added